fix max_int memory pool issue

+++ inc/Instance.h
@@ -5,9 +5,9 @@
 class Instance
 {
 public:
-	int num_of_cols;
-	int num_of_rows;
-	int map_size;
+	uint64_t num_of_cols;
+	uint64_t num_of_rows;
+	uint64_t map_size;
 
 	// enum valid_moves_t { NORTH, EAST, SOUTH, WEST, WAIT_MOVE, MOVE_COUNT };  // MOVE_COUNT is the enum's size
